app/requirements.txt
celery>=5,<6
django>=3,<4
django-debug-toolbar>=3,<4
django-extensions>=3,<4
gunicorn
mysqlclient
redis
requests
Werkzeug
